This commit was generated by cvs2svn to compensate for changes in r5809, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@5810 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/interrupt.h
@@ -241,8 +241,8 @@ _INT_FUNC void ack_dma(cpu_int_status_t *cs)
 _INT_FUNC void set_int(cpu_int_status_t *cs, int int_num,
 		       enum cpu_int value, CLOCK clk)
 {
-    set_nmi(cs, int_num, value & IK_NMI, clk);
-    set_irq(cs, int_num, value & IK_IRQ, clk);
+    set_nmi(cs, int_num, (int)(value & IK_NMI), clk);
+    set_irq(cs, int_num, (int)(value & IK_IRQ), clk);
 }
 
 /* ------------------------------------------------------------------------- */
